Fixing
Already up-to-date. Please enter the commit message for your changes. Lines starting
Already up-to-date. with 'Already up-to-date.' will be ignored, and an empty message aborts the commit.
Already up-to-date.
Already up-to-date. On branch master
Already up-to-date. Your branch is up-to-date with 'origin/master'.
Already up-to-date.
Already up-to-date. Changes to be committed:
Already up-to-date. 	modified:   Yasu/Yasu.munki.recipe